refactor(review): address xhigh code-review pass 3 findings on PR #731
Final review pass, three findings:

1. internal/linkgraph/wikilinks.go's isMarkdownName duplicated the
   exact ToLower-instead-of-EqualFold anti-pattern this PR already
   fixed, on the same feature's call path: crossfilereferenceintegrity's
   checkWikilinkAnchor (touched by this PR's earlier commits) resolves
   wikilinks through NewWikilinkIndex's fs.WalkDir callback, which
   called isMarkdownName once per workspace file — the same O(files)
   shape as the duplicatedcontent fix. Migrated to
   mdpath.IsMarkdownPath and removed the now-redundant local test (its
   cases are covered by internal/mdpath/mdpath_test.go).

   Declined as out of scope: a broader sweep of ~10 other unrelated
   hand-rolled .md/.markdown checks across internal/lint, internal/corpus,
   internal/lsp, internal/githooks, and internal/release — none of them
   sit on a call path this PR's 5 original fixes touch, and migrating
   all of them is a separate cleanup beyond "fix the top 5 issues the
   audit found."

2. fieldPatternCache (type + get/put + buildFieldPattern, added in the
   pass-2 commit) sat in the middle of the already-2600-line rule.go,
   unlike this package's existing convention of splitting cohesive
   concerns into their own files (runcache_wiring.go, scope_rules.go).
   Moved to fieldpatterncache.go / fieldpatterncache_test.go.

3. fieldPatternCache.get and .put were only exercised indirectly
   inside one combined test; split into TestFieldPatternCache_Get and
   TestFieldPatternCache_Put, each covering their nil-receiver,
   empty-cache, and populated-cache paths by name.

All fixes verified: go build, go vet, go test ./... (including
-race), golangci-lint (0 issues), gofmt, mdsmith check . (544 files,
0 failures).
